% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildMatrix.R
\name{buildMatrix}
\alias{buildMatrix}
\title{Gathers word frequencies into a matrix.}
\usage{
buildMatrix(dt, type = "docTerm", wordLimit = 5000, modelSize = 500,
  context = 5, method = "raw")
}
\arguments{
\item{dt}{A \code{docTexts} object that contains the full
                   text of each document in your corpus.}

\item{type}{A string (either "docTerm" or "wordContext") to
                   specify the type of matrix to be created. Document-
                   term matrices have columns that are documents.
                   Word-context matrices have columns that are keywords.}

\item{wordLimit}{A numeric value that limits the number of words (i.e.,
                   the number of rows) that will be counted. The default
                   is 5,000 words.}

\item{modelSize}{A numeric value, used for word-context matrices, that
                   sets the number of keywords to be evaluated. The default
                   value is 500, and the max is 5,000. Automatically, the
                   most frequently occuring words (not counting stopwords) are
                   used for analysis.}

\item{context}{A numeric value, for word-context matrices, that sets the
                   size of the context window. At 5 (default), the algorithm
                   will find words with 5 positions, before or after, each
                   of the keywords.}

\item{method}{A string, either 'raw' or 'proportional'. Determines with the
                   word frequencies will be reported as raw integers or in
                   proportion to the whole. Default is 'raw'.}
}
\value{
dm     a \code{docMatrix} object with frequency
               data for the corpus.
}
\description{
Gathers word frequencies into a matrix.
}
\examples{
dt = docTexts()
dm = getDocMatrix(dt)
dm = getDocMatrix(dt = dt, type = "docTerm", wordLimit = NULL, method = "proportional")
dm = getDocMatrix(dt = dt, type = "wordContext", wordLimit = 5000, modelSize = 500, context = 5, method = "raw")
}

