% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\title{Compose Vectors}
\usage{
compose(mat, positive = "", negative = "", operation = "*",
  fullResults = T, method = "cosine", margin = 1, threshold = 0)
}
\arguments{
\item{mat}{A term-document or word-context matrix (format either matrix or \code{docMatrix}).}

\item{positive}{A character vector of terms naming the rows to be aggregated.}

\item{operation}{A character vector (either "+" or "*"), that controls whether the vectors
will be added as simple sums or as products. Default is "*".}

\item{fullResults}{A logical value. If \code{TRUE}, the composite vector will be returned, without
computing similarity scores. If \code{FALSE}, only terms most similar to the
composite vector will be returned.}

\item{method}{The calculation to be performed ("cosine", "euclidean", "pearson", or
"covariance").}

\item{margin}{A numeric value. If \code{margin = 1} then rows will be computed. \strong{Note
that the current development version of this function only supported computing
over rows, so don't change this parameter.}}

\item{threshold}{A numeric value, between 0 and 100. See \code{\link{similarity}} for more 
details.}
}
\description{
Performs vector composition and negation, then returns the 12 most similar terms.
}
\examples{
data(eebo)
compose(mat = eebo, positive = c("wit", "sense"))
}

