% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eebo.R
\docType{data}
\name{eebo}
\alias{eebo}
\title{Data - EEBO word-context matrix}
\format{A matrix with 28,235 rows and 1,751 columns.}
\usage{
eebo
}
\description{
A word-context matrix drawn from the EEBO-TCP collection, built from all publicly
available files with publication years between 1640 and 1699.
}
\examples{
# Load the matrix, then analyze a term.
data(eebo)
find_similar(mat = eebo, keyword = "rights")
}
\keyword{datasets}

