% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_centroid.R
\name{find_centroid}
\alias{find_centroid}
\title{Find the centroid of word vectors}
\usage{
find_centroid(mat, positive = NULL, result = "analogies", sorted = T,
  exclude = T)
}
\arguments{
\item{positive}{A character vector of terms naming the rows to be computed.}

\item{result}{A string. If "analogies" then it returns similar terms, just like find_analogies.
                        if "vector" will return the centroid vector.}

\item{sorted}{A logical value.}

\item{exclude}{A logical value. If TRUE, the computed terms ('positive' and 'negative') will
              be excluded from the results.}
}
\description{
Find the centroid of word vectors
}
\section{Slots}{

\describe{
\item{\code{mat}}{A word-context matrix (format either matrix or \code{docMatrix}).}
}}
\examples{
data(eebo)
# 'arteries':'veins' for words that exist together in a common category
find_centroid(mat = eebo, positive = c("wit", "sense"))
}

