% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_context.R
\name{graph_context}
\alias{graph_context}
\title{Create a graph of similar words}
\usage{
graph_context(mat, keyword, use = "all")
}
\value{
A dendrogram, showing structure of 30 most-similar words.
}
\description{
Build dendrogram that represents the conceptual structure of a keyword.
}
\section{Slots}{

\describe{
\item{\code{mat}}{A word-context matrix.}

\item{\code{keyword}}{A string.}

\item{\code{use}}{A string ('para', 'syn', or 'sig'.) See \code{\link{find_similar}}.}
}}
\section{What it does}{

This function runs \code{find_similar} over a word-context matrix and looks for the
thirty most similar terms. After finding the thirty most similar terms, it clusters
them, performing a simple but effective word-sense disambiguation (WSD).
}

