% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locke.R
\docType{data}
\name{locke}
\alias{locke}
\title{Data - Locke's Treatise (word-context matrix)}
\format{A matrix with 5,000 rows and 500 columns.}
\usage{
locke
}
\description{
A word-context matrix drawn from the EEBO-TCP edition of John Locke's
Two Treatises of Government (TCP# A48901.)
}
\examples{
# Load the matrix, then analyze a term.
data(locke)
graph_context(mat = locke, keyword = "woman")
}
\keyword{datasets}

