% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nlaw.R
\docType{data}
\name{nlaw}
\alias{nlaw}
\title{Data - 17th Cent. Political Theory (word-context matrix)}
\format{A matrix with 5,000 rows and 500 columns.}
\usage{
nlaw
}
\description{
A \code{docMatrix} object containing a word-context matrix for a sample dataset
of 33 17th-century documents relating loosely to political theory.
}
\examples{
# Load the matrix, then analyze a term.
data(nlaw)
graph_context(mat = nlaw, keyword = "laws")
}
\keyword{datasets}

