% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector in a matrix}
\usage{
normalize(x, margin = 1)
}
\arguments{
\item{margin}{The direction of comparison. If "1", normalization will be computed
for a row. If "2", normalization will be computed over a column.}

\item{mat}{A word-context matrix}

\item{term}{The word you are evaluating}
}
\description{
Return normalized vector in a word-context or term-document matrix
}

