% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{sig}
\alias{sig}
\title{Return the signatures from a matrix}
\usage{
sig(mat, keyword, margin = 1, fullResults = F)
}
\arguments{
\item{mat}{A term-document or word-context matrix.}

\item{keyword}{The term you are evaluating, either a keyword or document 
name.}

\item{margin}{Numeric value: 1 or 2. If 1, calculations are performed over the rows. If 2, over
the columns.}

\item{fullResults}{Logical value. Default is false.}
}
\value{
If \code{fullResults} is true, all results are included in a full-length vector. If
        false, only the 12 top words will display.
}
\description{
Prints the sorted values from a term-document or word-context matrix
}
\section{What it does}{

Sorts and prints the most frequent collocates for words and documents. In a term-
document matrix, it will return either the word counts for a document or the top
documents for a given word.

In a word-context matrix, the relationship is similar. Selecting \code{margin = 2}
will find the keyword's column, then return the word counts of the concordance for
that term (assuming that it was among the keywords used to build the matrix). 
Selecting \code{margin = 1} will return the keyword's row, showing in which
concordances it had the highest representation.

For comparison, see the example of 'rights' given below.
}
\examples{
# For most frequent collocates in a word-context matrix
data(eebo)
sig(mat = eebo, keyword = "rights", margin = 1)
sig(eebo, "rights", margin = 2)

# For most frequent words in a document
data(shakespeare)
sig(mat = shakespeare, vec = "TN", margin = 2)
sig(shakespeare, "wife", margin = 1)

}

