% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_dendrogram.R
\name{similarity_dendrogram}
\alias{similarity_dendrogram}
\title{Create a graph of similar words}
\usage{
similarity_dendrogram(mat, keyword, method = "cosine", margin = 1,
  threshold = 50, numResults = 30, numGrps = 5)
}
\arguments{
\item{mat}{A word-context matrix.}

\item{keyword}{A string.}

\item{method}{A character string: 'cosine', 'euclidean', 'pearson' or 'covariance', which names
the mathematical similarity test to be performed. Default is 'cosine'.}

\item{margin}{Numeric value: 1 or 2. If 1, calculations are performed over the rows. If 2, over
the columns.}

\item{threshold}{Numeric value: 0 to 100. Default is 50.}

\item{numResults}{Numeric value. The number of words to be displayed in the 
graph. Default is 40.}

\item{numGrps}{Numeric value. The number of groups in which you'd like to divide
the display. Default is 5.}
}
\value{
A scatterplot showing structure of 30 most-similar words.
}
\description{
Build a scatterplot that represents the conceptual structure of a keyword.
}
\section{What it does}{

This function runs \code{\link{similarity}} over a word-context matrix and looks for the
thirty most similar terms, then clusters them.
}

