% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Return z-score for a keyword}
\usage{
zscore(mat, term, margin = 1)
}
\description{
Return z-score for a keyword in a word-context matrix
}
\section{Slots}{

\describe{
\item{\code{mat}}{A word-context matrix}

\item{\code{term}}{The word you are evaluating}

\item{\code{margin}}{The direction of comparison. If "1", zscore will be computed
             for a row. If "2", zscore will be computed over a column.}
}}

